[Streamer] Move player implementations to cpp files
